Tests migration.
