+++ cardano-offchain-stableswap/src/lib.rs
@@ -1,3 +1,2 @@
 mod stable_swap_amm_actions;
 pub mod stable_swap_invariant;
-mod test;
